PAD: insert DLS regularly while SLS transmission
When a slide is transmitted, now every 50 PADs the current DLS is inserted
(and thereby reread from file). Hence in the best case (AAC-LC @ 48 kHz), DLS
is inserted every 1000ms; in the worst case (HE-AAC @ 32 kHz) it is inserted
every 3000ms. This way a listener will get DLS much earlier after switching to
a service, compared to the previous situation where the slide transmission was
not interrupted for DLS insertion.

Note that there still remains a delay without any PAD, between the end of a
slide transmission and the start of the next one's transmission.

Also note that adding a feedback channel from the audio encoder to ODR-PadEnc
will make this improvement obsolete.