Add recovery on broken pipe
When the opposite side of the PAD FIFO disconnects (e.g. due to crash of
the audio encoder), ODR-PadEnc crashed as well due to SIGPIPE.

This commit changes the behaviour. From now on instead error messages
are displayed and it is possible for a new instance of the audio encoder
to reconnect to the still running ODR-PadEnc instance.