Add uniform PAD encoder
The uniform PAD encoder encodes every PAD on demand and is kind of
synchronized to the audio encoder by means of the frame duration.

Insertion intervals for slides and labels can be set indepentent from
each other. It is also possible to specify how often the label is
inserted.

If the slides interval "0" is used, the next slide is inserted just
after the previous one has been transmitted. This is useful e.g. for
stations that transmit just a logo slide.

An initial burst count can be set to ensure that an audio encoder has
enough PADs available e.g. in case the audio encoder encodes DAB+
Superframes at once and hence needs all related PADs.

If a slide/label is still in transmission when the transmission of the
next one is scheduled, the new transmission is skipped and a warning is
shown. In this case it makes sense to increase the PAD length or to
instead decrease the slide size or label insertion interval (`-L`).

This new PAD encoder does not require any changes on the audio encoder
side. Only in case of MP2, a recent revision of ODR-AudioEnc has to be
used (commit ce25f2c or later), as it fixes a problem with PADs that
solely consist of the F-PAD.