DLS: add support to request DLS file re-read
A re-read of any of the used DLS files can now be requested by storing a
file with the same filename which is suffixed by `.REQUEST_DLS_REREAD`.

When the next PAD is going to be encoded and such a file is present, the
encoder changes the current DLS file to the related DLS file, if needed
(only applies if multiple DLS files are used). The DLS file is re-read
and the request file is deleted. The DLS is also immediately inserted
into the PAD transmission.